refactor: cleaning up index.ts
